Refactor library ID mapping methods to use external IDs and update requirements for Jellyfin SDK
